+++ src/urldata.c
@@ -243,13 +243,13 @@ urldata_search_pexels (const char *s_query,
 }
 /*----------------------------------------------------------------------------*/
 /**
- * @brief  Search Pixbay for images.
+ * @brief  Search Pixabay for images.
  */
 UrlData *
-urldata_search_pixbay (const char *s_query,
-                       const char *s_search_opts,
-                       const char *s_api_key,
-                       const int   i_page)
+urldata_search_pixabay (const char *s_query,
+                        const char *s_search_opts,
+                        const char *s_api_key,
+                        const int   i_page)
 {
     UrlData  *ud_data = nullptr;
     char     *s_url   = nullptr;