+++ src/webpixabay.h
@@ -1,5 +1,5 @@
 /**
- * @file  webpixbay.h
+ * @file  webpixabay.h
  * @copyright Copyright (C) 2019-2026 Michał Bąbik
  *
  * This file is part of Wall Changer.
@@ -17,33 +17,33 @@
  * You should have received a copy of the GNU General Public License
  * along with Wall Changer.  If not, see <https://www.gnu.org/licenses/>.
  *
- * @brief  Settings for searching the Pixbay website.
+ * @brief  Settings for searching the Pixabay website.
  *
  * @author Michal Babik <michal.babik@protonmail.com>
  */
-#ifndef WEBPIXBAY_H
-#define WEBPIXBAY_H
+#ifndef WEBPIXABAY_H
+#define WEBPIXABAY_H
 
 #include "nstrings.h"
 #include "webwidget_s.h"
 /*----------------------------------------------------------------------------*/
 /**
- * @brief  Search in Pixbay database.
+ * @brief  Search in Pixabay database.
  *
  * @param[in,out] ww_widget  WebWidget with widgets and search data
  * @param[in]     ns_data    N strings with API keys
  * @return        none
  */
-void   pixbay_search              (WebWidget      *ww_widget,
-                                   const NStrings *ns_data);
+void   pixabay_search              (WebWidget      *ww_widget,
+                                    const NStrings *ns_data);
 /*----------------------------------------------------------------------------*/
 /**
- * @brief  Dialog with Pixbay service settings.
+ * @brief  Dialog with Pixabay service settings.
  *
  * @param[in,out] ns_data  N strings to get and write API keys
  * @return        Dialog response
  */
-int    pixbay_settings_dialog     (NStrings       *ns_data);
+int    pixabay_settings_dialog     (NStrings       *ns_data);
 /*----------------------------------------------------------------------------*/
 /**
  * @brief  Options for image search dialog.
@@ -52,7 +52,7 @@ int    pixbay_settings_dialog     (NStrings       *ns_data);
  * @return        String with image options. After use it should be freed
  *                using free.
  */
-char * pixbay_search_opts_dialog  (WebWidget      *ww_widget);
+char * pixabay_search_opts_dialog  (WebWidget      *ww_widget);
 /*----------------------------------------------------------------------------*/
 #endif